searching map and make fragment
