+++ app/src/main/res/layout/activity_main.xml
@@ -4,32 +4,11 @@
     android:layout_height="fill_parent"
     android:orientation="vertical" >
 
-    <LinearLayout
+    <fragment
+        android:id="@+id/mainfragment"
+        android:name="com.aplusstory.tmap.MapFragment"
         android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:orientation="horizontal">
+        android:layout_height="match_parent" />
 
-        <Button
-            android:id="@+id/bt_findadd"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="검색" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/mapview"
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@android:color/white"
-        android:layout_weight="1" >
-    </LinearLayout>
-    <FrameLayout
-        android:id = "@+id/fragment_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        >
-    </FrameLayout>
 
 </LinearLayout>
