+++ app/src/main/res/layout/fragment_map.xml
@@ -6,9 +6,17 @@
     tools:context=".MapFragment">
 
     <!-- TODO: Update blank fragment layout -->
-    <TextView
+    <Button
+        android:id="@+id/bt_findadd"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="검색" />
+
+    <LinearLayout
+        android:id="@+id/mapview"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
-
+        android:layout_weight="1"
+        android:background="@android:color/white"
+        android:orientation="vertical"></LinearLayout>
 </FrameLayout>
